backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
httpx==0.25.1
pytest-asyncio==0.21.1
aiosqlite==0.19.0
greenlet==3.0.1
asyncpg==0.28.0
psycopg2-binary==2.9.9 